implements exception handler

+++ Projetos/dscommerce/src/main/java/com/agvsistemas/dscommerce/services/ProductService.java
@@ -2,13 +2,16 @@ package com.agvsistemas.dscommerce.services;
 
 import com.agvsistemas.dscommerce.dto.ProductDTO;
 import com.agvsistemas.dscommerce.entities.Product;
+import com.agvsistemas.dscommerce.exceptions.DatabaseException;
 import com.agvsistemas.dscommerce.exceptions.ResourceNotFoundException;
 import com.agvsistemas.dscommerce.repositories.ProductRepository;
-import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class ProductService {
@@ -16,19 +19,17 @@ public class ProductService {
     @Autowired
     private ProductRepository productRepository;
 
-    @Transactional
+    @Transactional(readOnly = true)
     public Page<ProductDTO> findALl(Pageable pageable) {
         var entityList = productRepository.findAll(pageable);
         return entityList.map(ProductDTO::new);
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        var entity = productRepository.findById(id);
-        if (entity.isEmpty()) {
-            throw new ResourceNotFoundException("Product #" + id + " not found!");
-        }
-        return new ProductDTO(entity.get());
+        var entity = productRepository.findById(id)
+            .orElseThrow(() -> new ResourceNotFoundException("Product not found!"));
+        return new ProductDTO(entity);
     }
 
     @Transactional
@@ -47,9 +48,17 @@ public class ProductService {
         return new ProductDTO(entity);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException();
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     public void copyDtoToEntity(ProductDTO dto, Product entity) {
